flask_virtual_hospital/requirements.txt
Flask==3.1.1
Werkzeug>=3.0.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
python-dotenv==1.1.0
pdfplumber==0.11.6
Pillow>=12.0.0
